CTS related utilitiy package
This adds the tsuite package which is used by the utility script to
run an apk on Firebase Testlab.
(see https://skia-review.googlesource.com/c/skia/+/89163)

This is in the buildbot repo, because eventually code will be
added to ingest the results of the testruns that that way we keep
related code in the same repo.

Bug: skia:
Change-Id: Idbfdd52013de0f80bc13370e2a72c2884f0363aa
Reviewed-on: https://skia-review.googlesource.com/90264
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Stephan Altmueller <stephana@google.com>