Update to polymer 1.10+, addressing HTML Imports deprecation
One thing that we were doing wrong that only manifested itself in the
new version of Polymer was over-using custom-style, which messed up many
CSS mixins (specifically, when trying to use them).

This CL removes custom-style where it is not needed and adds an
explanation to where it is needed, as well as a word of caution,
so custom-style does not re-propogate.

Bug: skia:
Change-Id: Iee9835479db19b4624c36b9a4d644ea6bf309e89
Reviewed-on: https://skia-review.googlesource.com/90483
Reviewed-by: Eric Boren <borenet@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Stephan Altmueller <stephana@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>