Remove fix-boot-order.yml.
This was required for UEFI boot because Windows changes the UEFI boot
order. We have switched to legacy (BIOS/MBR) boot, for which this
Ansible script does not work. (bcdedit is not able to change non-UEFI
boot options, which also means that Windows is not able to change the
boot order.)

Change-Id: I600651123a7e45ea30a2f89c6b90636dfe31aad7
Reviewed-on: https://skia-review.googlesource.com/109383
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Ben Wagner <benjaminwagner@google.com>