Use jsdoc to document our pure JS modules and custom elements.
Documenting custom elements involved writing a jsdoc plugin
to support @evt and @attr tags to document event and attributes
respectively.

The rest of the CL is adding/converting documentation
on some of the common files to show how comments will
work.

Right now the docs go in the 'out' directory and are not
checked in, pending a decision on where we would like to host
those files.

Screenshots:
  https://screenshot.googleplex.com/jj11wY0pMpO.png
  https://screenshot.googleplex.com/jvaxvoBYncZ.png

Bug: skia:
Change-Id: I3bf9f56469d95618a30110eddbaa48e3c93987df
Reviewed-on: https://skia-review.googlesource.com/112567
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>