[pushk] A new version of the pushcli command line tool for the kubernetes world.
Instead of pushcli which modifies a json doc we store in GCE, the kubernetes
files will be checked into our repo. This tool makes it easier to update
those files to the latest version of an app.

Note that you can always manually edit a yaml file to point to the version
of an image that you want to deploy.

Bug: skia:
Change-Id: I6baacd4e9ae367746d1627e595bd01a297826a12
Reviewed-on: https://skia-review.googlesource.com/123743
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Ben Wagner <benjaminwagner@google.com>
Auto-Submit: Joe Gregorio <jcgregorio@google.com>