[fuzzer] Port backend to Kubernetes.
This yak shave was prompted by the fact we are running a very old version
of Clang and that recreating a fuzzer instance was kind of a pain because
of the setup/startup tensions that sometimes happened.

And it was time.

Basic steps:
1. Port setup.sh and startup.sh to the Dockerfile, building off the ideas
in the skia-release Dockerfile.
2. Create a service token (create-fuzzer-be-sa.sh). Whitelist this token
in the old project's bucket auth.
3. Make kubernetes yaml files in skia-public-config [1]
4. Take the arguments from the .service files and put the shared ones
in the Dockerfile, and the unique ones in the appropriate .yaml files
5. Make sure auth works:
  5a. Change auth from auth.NewDefaultJWTServiceAccountClient to
auth.NewDefaultTokenSource.
  5b. Remove common.CloudLoggingOpt() from InitWithMust
6. Misc Docker/K8s related changes - like tweaking dependencies on
os.Hostname().


[1] https://skia.googlesource.com/skia-public-config/+/a5911039243ffbd8af08558152d5fd4cfaf3e796
Bug: skia:
Change-Id: Ib1dcc316868d1d4c3ecf70878de3ee02f5069f81
Reviewed-on: https://skia-review.googlesource.com/138262
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>