[task scheduler] Use priority and dependencies in scoring.
Previously the TaskSpec priority was essentially ignored, since Task
Scheduler would only use its internal scoring to decide when to trigger
tasks. In this CL, Swarming priority is fixed and we compute task scores
based on the jobs that depend on the task. The formula is
  1 - (1-<job1 priority>)(1-<job2 priority>)...(1-<jobN priority>)
so a task priority increases as the number of dependent jobs increases.

Bug: skia:8145
Change-Id: I91b70ad51d3948b5d312ccd24ceeb5af0c0b39d3
Reviewed-on: https://skia-review.googlesource.com/141564
Commit-Queue: Eric Boren <borenet@google.com>
Reviewed-by: Eric Boren <borenet@google.com>