Refactor to fix elusive race condition in ingestion
Refactors that should help towards fixing a hard to
reproduce race condition in TestPollingIngesterWithStore.
The few times I was able to reproduce the issue, it seemed
to be related to BT emulator being in an undesired state.
This removes all tables before it each run hopefully eliminating
the issue.

Bug: skia:8528
Change-Id: I12a5321ab9338d0094b233fff11af607f236e493
Reviewed-on: https://skia-review.googlesource.com/c/169461
Commit-Queue: Stephan Altmueller <stephana@google.com>
Reviewed-by: Ben Wagner <benjaminwagner@google.com>