[task scheduler] PubSub handler: check the pendingInsert map before the DB
There was a race condition in which we could attempt to retrieve the task
from the DB, find that it doesn't exist, and while we're waiting for the
pendingInsert mutex to be acquired, it is inserted into the DB and removed
from the pendingInsert map.  Thus it appeared that the task was not in the
DB or in the pendingInsert map.  By checking pendingInsert first, we avoid
this problem and eliminate a DB lookup.

Bug: skia:
Change-Id: I91fbe3064989b112270216dc7ec8514cee4fbd37
Reviewed-on: https://skia-review.googlesource.com/c/188764
Commit-Queue: Eric Boren <borenet@google.com>
Reviewed-by: Ben Wagner <benjaminwagner@google.com>