perf - call gitinfo.Update() deliberately.
gitinfo.Update() used to be called regularly as a side-effect of clustering, but
with the new, faster way of clustering by querying based on tile size and not
on time/date the calls that triggered gitinfo.Update got dropped. Solution is to
add gitinfo.Update() calls into the dataframe refresher.

Bug: skia:
Change-Id: I68b68866eae006e31256a120fdf384338a423c0f
Reviewed-on: https://skia-review.googlesource.com/c/196041
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>