[gold] add service account to authorized users list
Removes metadata as source of authorized users (which I believe
was blank anyway?)

Add a few fixes to goldctl to properly use the credentials.

Along the way, I thought there might be a bug in login, so
I added some tests to try to catch it.  Bug was elsewhere,
but the tests may still be useful.

Change-Id: If2c95779849cfd5af7c7e437d7d6ff3ef89e09e5
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/208504
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Ben Wagner aka dogben <benjaminwagner@google.com>