[gold] Refactor monolithic storage
In anticipation of adding in the new tracestore interface.

Removes sampler/main.go because it's not being used actively
and the bitrot is getting annoying to maintain.

Note to the reviewer: you can ignore most of tilesource.* because
it will get a large make over when the bt_tracestore is
adopted (namely, much of tilesource will go away).


Bug: skia:9096
Change-Id: Id78c4d8be4c67ff4098549ea0d3064ee7c57ab29
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/219477
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Ben Wagner aka dogben <benjaminwagner@google.com>