Ignore dead and quarantined bots for bot check.
Dead and quarantined bots tend to keep the dimensions that they had
when they stopped working. Branch dimensions that only match dead and
quarantined bots are probably out of date.

Even if the dimensions are correct, it's probably useful to get an
alert when there are no working bots for a task on the CQ.

Change-Id: I57b032066726001168889a0c47c98f5230eace80
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/224640
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Ben Wagner aka dogben <benjaminwagner@google.com>