[gold] Only store the latest options per trace/tile
In practice, storing all options per tile led to a massive slowdown
on GetTile because the GetRow had to allocate memory for all those
bytes, even though most got ignored.

Ingestion will change to call Put with the timestamp of the commit
(instead of timestamp of the file, which was the original plan).

This also adds a few other options-related optimizations, like
interning and avoiding regexp.

Bug: skia:9096
Change-Id: I99a8ae94650e6fae491b09a14a52c3429e749292
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/226018
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Ben Wagner aka dogben <benjaminwagner@google.com>