Fix for infra-sk location.
It turns out that the default location to resolve modules is "node_modules",
i.e.  a relative directory, so we have to over-ride that with an absolute
path.

Also, each project that uses infra-sk must add in the deps that
infra-sk needs, i.e. common-sk, elements-sk, and lit-html.

Change-Id: I5154b86ca8efe22759c5f607e3e90c7ca8f42750
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/226520
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>