[gold] Change "issue" -> "ChangeList" and make the ID a string.
This affects ExpectationsStore and the surrounding code.

I realize that this may cause some issues with existing
CLExpectations, but I checked and there are none (which makes
sense, since viewing TryJob results has been broken/slow for
a while now). So I'm ok making a breaking change to the
schema.

This also adds tests for Triaging, which 1) is untested and 2)
changes a result of the parent change.

This requires creating new Firestore collections for
expstore_*  I added a v2 to these and a migrator (again, making
the simplifying assumption that there are no outstanding
CLExpectations)

Bug: skia:9340
Change-Id: I5aab7029290029aded7da9444acf7cf4449a6ba3
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/242082
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Ben Wagner aka dogben <benjaminwagner@google.com>