[gold] remove rate-limiting from known hashes
The revert of flutter's presubmit testing was likely due to rate limiting
on the KnownHashes endpoint. https://github.com/flutter/flutter/pull/43227

This CL adds a TTL cache for that endpoint and removes the rate limiting,
so that the baseliners can serve it at a higher load.

Change-Id: I018c3db0ce9a2396af2090898e048fad8ed86bb0
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/249890
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Ben Wagner aka dogben <benjaminwagner@google.com>