[gold] Run byblame and search in parallel
This is much easier to do with the slices of traces from earlier CLs.

When the pod isn't busy (and has CPUs to spare), I see byblame be as
low as 80ms for the images corpus (was averaging 1.8s previously).

The search endpoint saw some more modest improvements (about 50%
faster for a few searches I tried), but the search endpoint latency
is dominated by the Polymer side rendering it (or the diffserver
computing the closest diff).

Bug: skia:9080
Change-Id: I80cb0632b65b459e1de18e224952327160f175f0
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/254416
Reviewed-by: Ben Wagner aka dogben <benjaminwagner@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>