[gold] Reduce memory by not locking SearchIndex in warmer closure
After adding in the pre-slice maps, we were accidentally
keeping around all the traces in that map because
the search index was locked away in a closure when warming
Since warming takes 15 minutes or so and we refreshed a tile
after 5 minutes, we were keeping an extra 3-4 traces around.

This appears to help calm down the memory usage again.
I spent some time looking at bt_tracestore again, because
the memory blame blames things in loadEncodedTraces that
I would have thought would have been GC'd, but I now believe
that go does something clever when copying things around
that might make it look like they are still there but are not.

Additionally, we were incorrectly keeping the row keys around,
which can be prevented with a copy, just like in Perf.

Change-Id: I0e913330c10ddccc3cea4d794b2d447a8eb92ea1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/254800
Reviewed-by: Ben Wagner aka dogben <benjaminwagner@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>