[gold] Consolidate text images into test package
I'm beginning to like this pattern of having some
test data in a given package that is shared by multiple
other packages. This means we can avoid some duplication
and have a consistent "story" in place, which makes it
easier for me to reason about test expectations because
I don't have to "relearn" a test corpus for each package
I visit.

Change-Id: Idd1167de2b6c59f763739cf976916ba846d202c7
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/256366
Reviewed-by: Ben Wagner aka dogben <benjaminwagner@google.com>
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>