[gold] Make puppeteer_test.js flaky to test out the new Infra-PerCommit-Puppeteer tryjob.
This CL contains a test that I've been using during development to test out the recipe with trybots. Let's land this CL to see if it works with normal bots as well.

I'll revert this CL once we're convinced that things work as expected.

Bug: skia:9608
Change-Id: Ic4ca60a18357a657fe4ad67d31fc6d9644c72670
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/259357
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>