[perf] Fix percent, absolute, and cohen step detection.
A radius of N has traditionally meant we constructed
a trace of length 2*N+1 around the commit of interest,
but that means the two populations we are checking are
of different sizes. That is before the commit = len N.
The half after the commit is of len N+1 since it also
contains the commit.

This is particularly problematic for absolute and percent
Alerts where the desired radius is 1. See the associated
bug for more details.

Also update the stepfit_test.go to conform to the expected
test structure.

Bug: skia:9921
Change-Id: I1f623eb3f9d399c8a1d0c83bcd9da9c0928b4bb9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/271856
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>