A few fixes to get hotspare working again
The lab network is possibly slower than previously because cloud logging
kept hitting the dial timeout. Increased the timeout to 500ms.

Fix the 'ip address *add*' command.

Passing "localhost," as a literal argument to ansible-playbook caused it
to want to run the playbook on hosts named `"localhost` and `"`.

It seems 'ifconfig' assumed the VIP should have a mask of 24, whereas
'ip address add' assumes it's 32. I think this only matters for
consistency with the master, and so that documented commands work.

Change-Id: I66252cd5cccf3ccdb7939b4cd0c7a20a6a8c5df2
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/273397
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Ben Wagner aka dogben <benjaminwagner@google.com>