Cancel the context in cleanup for Firestore testing clients
This also adds the context as a param, so the same context that is used
in a test can be passed into the test client. (It felt odd to have
two different canceled contexts)

I'm hoping this helps fix some strangeness in firestore tests that I
am seeing.

Change-Id: Ie705e702e1bbb1dd58ff3d494948801d74d2bd58
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/273994
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>