[goldctl] CloudClient: Skeleton code to support non-exact image matching algorithms.
This CL is essentially a refactor around the logic in CloudClient#Check() and CloudClient#addTest() that checks whether the given image is labeled as positive.

Basically, said logic now resides in CloudClient#matchImageAgainstBaseline(), which is shared by both the Check() and addTest() methods. In addition, matchImageAgainstBaseline() parses the optional keys and tries to instantiate the specified image matching algorithm, defaulting to exact matching if none is specified to preserve backwards compatibility.

This change should be a no-op as long as no image matching algorithm is specified.

Bug: skia:9527
Change-Id: I2b7ad2b438734fa8d1d49e8b2d0149812925a716
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/276280
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>