[power] Rename to Controller and introduce DeviceID
This makes the package align with Go style a bit better. No functional
changes intended.

Note to the reviewer: start in powercycle.go for the renames to
interfaces and the new types, then look at the rest in any order.

I did not do much to edgeswitch_client because I have larger plans
for that (including combining it with edgeswitch_group) and did not
want to complicate the diff.

Bug: skia:9673
Change-Id: I25250f21b841939b097a1f93523af48cc7167488
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/283502
Reviewed-by: Weston Tracey <westont@google.com>
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Eric Boren <borenet@google.com>