[infra] More util cleanup. Move some files around.
concurrency.go was unused (formerly used in Gold?).

repo.go was only used by vcsinfo/testutils, so it was moved
and unexported.

context.go was used in one package (go/exec), so it was moved
there and unexported.

zip.go was used in a few places, but seemed like it could
cleanly get moved into a subpackage. Since new package was
named zip, zip.ZipIt seemed clunky, so a sly rename to
zip.Directory made me feel better about it.

Change-Id: I9e150e5a5130dbf096ca47c4a6c47ac679715106
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/286218
Reviewed-by: Eric Boren <borenet@google.com>