[update go deps] Fall back to "go get" individual packages
This is currently failing because one module has no .go files at the
root level. It works if we use a package from the module as input to
"go get", but doing so all the time has other problems (eg. renamed
or deleted packages cause spurious failures), so we try the module
first, and if that fails we'll try packages until one succeeds or we
run out of packages to try.

Change-Id: I0dba23d0b1ccacd82c8b50ba6638db4d09cc59e1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/287821
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Eric Boren <borenet@google.com>