[gold] Remove commenting logic from ingestion
Comments are noisy and having the logic to decide to comment be
(essentially) at the point of ingestion makes it hard to play around
with that data. This logic will be deduplicated with the logic in the
search API, i.e. UntriagedUnignoredTryJobExclusiveDigests.

This is now possible because of the CL indices that make the search
API fast and not overload the Firestore backend.

Bug: skia:10218
Change-Id: Ia21f68e35e43a9b279e626f47401b3ac3783d570
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/287880
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>