[gold] Make commenter use searchAPI
This attempts to reduce the false positive rate of commenter sending
comments when an untriaged digest was first seen on master branch. It
does this by leveraging the logic in the search API instead of trying
to do its own thing. The search API is hopefully the freshest data
with respect to digests seen on master. Additionally, this makes commenter
less likely to bug somebody about an untriaged digest that has already
been triaged.

Using the power of the changelist index, this also lets us estimate to
the user how many untriaged digests there are on their CL.

Suggested review order:
 - code_review/types.go
 - fs_clstore.go
 - commenter.go
 - commenter_test.go
 - Everything else

Bug: skia:10218
Change-Id: Ifa8f368c63f725e5d8d1ca474633be3860335bf1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/287883
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>