[gold] Copy karma.common.js file from common-sk.
This file was copied verbatim from the common-sk directory.

We will make any necessary changes to this file to support TypeScript tests.

Eventually this file will be moved to //infra-sk to make it available to other apps in the Skia Infra repo, and we will remove the karma.common.js file in the common-sk repository.

Bug: skia:10246
Change-Id: Ibde57f58f48d686bfb712131e9237d59605c7203
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/289619
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>