[gold] Remove race condition between CL ingestion and indexing
I believe the primary fix is in indexer, by updating the time and the
CL at the end of ingestion (after the TryJobResults have been stored).
This was exacerbated by the indexer updating the timestamps when it
really shouldn't have.

As an added measure of precaution, this adds a grace period to the
creating of the CL index so as to be more forgiving if there somehow
is some lag between ingestion and the data being ready (unlikely, given
that Firestore should be write-consistent).

Bug: skia:10265
Change-Id: I2acca3b8721f5d0915de47adfe61ca18f64f088b
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/290919
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>