[gold] Add new RPC for rewritten /list page
As we rewrite the digests by list page (found at /list) in
lit-html, I want to take the opportunity to simplify the page.

If we have the page focus on telling the user "these are the
tests and here are how many unique digests each test produces",
then things get a lot simpler. For example the blame column
(which I will confess I don't know how to interpret) can be
removed as well as the data needed to render it.

As such, the API that serves that page has been rewritten
with the bare minimum data. This API is also tested.

Bug: skia:9525
Change-Id: I37bee3972a77034aeb1db5cd31cb8f49c2748f7e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/296396
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>