[gold] Add cluster-digests-sk element.
This is not the final, complete element, but is a starter
to address some boilerplate with respect to adding in d3
and adding some color CSS variables.

Future CLs will make this more interactive and may need
to change it as it gets integrated with the ported page,
but this should be a manageable CL for initial review.

Note to the reviewer: If you are not familiar with d3,
I'm happy to take some time and step through this with you.

Change-Id: Ic36b91eb6bf3ffcfee44e6ad499f60b0aef92165
Bug: skia:9525
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/299861
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>