[infra-sk] Move the fast filter on query-sk.
I got a bug report that fast filter should work on both params and values, when
it already does that. So it appears that it isn't clear that is what filtering
does.

This CL moves the filter to below both params and values and adds placeholder
text that indicates what the control does.

Bug: skia:10327
Change-Id: I4c14e2a7e00b4db9ceced1df1db82b6bb26b4800
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/301842
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>