Remove skiaversion
It was helpful in the past, when it wasn't trivial to determine at what
version of the code the executable was built, but we have much more
visibility and control now with kubernetes. I have not needed the log
lines or version link in recent memory. Additionally, having a
constantly-updating version embedded in the executable prevents our
builds from being reproducible.

Change-Id: I3341adbfd7b8e6922fdc0e4ddc8e85fbfe514f3b
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/302857
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Eric Boren <borenet@google.com>