[infra-sk] Bazel macro css_to_scss().
This is used to turn the plain CSS files (*.css) in the elements-sk NPM package into Sass files (*.scss) that can be used with the sass_library() and sass_binary() Bazel rules.

Additionally, an elements-sk_scss sass_library() is defined to demonstrate the css_to_scss() macro. Said sass_library() is used with the theme-chooser-sk element in a follow-up CL.

Bug: skia:10614
Change-Id: Ibaf865bf7aa823cd85d1cfdbc37dfe436936d989
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/312399
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>