[gold] Change location for tryjob ingestion
I had some difficulty when diagnosing skbug.com/10723 trying to determine
where data for a given CL was going in GCS. This should make that path
structure more clear by using the CL (and PS) id instead of the githash.

This also shortens the timestamp out of the path because it is redundant
with the name of the file.

Change-Id: Ia8d5a1ffd8b2fb70dad92daaa2874c01744eb1cb
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/317559
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>