[status] Add dialog-sk element.
This adds a dialog element for use on the tasks table.
It is a single element of the form that is essentially:

{{TitleSection}} {{action button}}
---
{{DetailSection}}
---
{{comments-sk}}

These sections are filled in based on task, taskspec, or commit data.

In additional to the element there is some minor cleanup:
--make comments-sk more robust for use in parent elements
--refactor task-driver-sk test data for use in details-dialog-sk tests
--add settings injection missing from commits-table-sk-demo

Change-Id: If20a021e975988abce3e96c41046afb9dc607036
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/320722
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Weston Tracey <westont@google.com>