[bazel] Custom RBE toolchain container with extra deps.
This CL extends the default RBE toolchain container with a Chrome installation, which is required to run Karma and Puppeteer tests. The set of packages installed is exactly the same as those in //puppeteer-tests/docker/Dockerfile.

In the future, we will include the Cloud Firestore and Cloud BigTable emulators in the toolchain container as well, which are required to run some of our Go tests.

To build the Docker container:

    $ bazel build //:rbe_container_skia_infra

To push the container to GCR:

    $ bazel run //:push_rbe_container_skia_infra

Reference documentation:

- https://cloud.google.com/remote-build-execution/docs/create-custom-toolchain

- https://github.com/bazelbuild/rules_docker

Bug: skia:10614
Change-Id: Id38713bfc9bec9c82aa9a0ba38c736f42f108e29
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/334317
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Eric Boren <borenet@google.com>