[perf] Fix mutation of global ParamSet.
Found this error when monitoring of ParamSets began. The Android ParamSets would
oscillate between two values:

  http://go/oscillating-paramsets

It turns out the psrefresher was starting with the most recent tile ParamSet
and then adding in the ParamSet of the previous Tile, but not making
a copy of the original Param, thus modifying the ParamSet in sqltracestore's cache.

A better longer-term fix will be to distinguish between normal and read-only ParamSets.

Change-Id: I6ef7eb0cfc40509844fef575b6004e9296fa1b01
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/337403
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>