Find image uses in layers, cycler-button emits event.
The image resource viewer previously reported images as unused if their only use was in a layer update.
this change finds those uses and puts them in a new table. The jump feature works on them and opens the layer update inspector.

cycler-button now operates via an event as requested in earlier cls.

A DefaultMap class is added and used throughout because that's how I was using Map most of the time
works just like collections.defaultdict in python

Change-Id: I76fa501b04643b69955181045103992450d22546
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/337539
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Nathaniel Nifong <nifong@google.com>