[gold] Add ToSQL model for writing tables directly to SQL
This adds some important code for bulk inserts -
sql.ValuesPlaceholders
This allows us to insert multiple items at once, but we do not
have to convert them to strings ourselves, but instead can
the SQL driver to do that for us. This has benefits including
performance and resistance to SQL injections.

Bug: skia:10582
Change-Id: Ide55695a4a20cd18faf2fa0e5dfafa19f6c11692
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/348784
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>