[bazel] Tag //:rbe_container_skia_infra as "no-remote" because it cannot be built on RBE.
This is because the container_run_and_commit rule requires the Docker daemon to be running, which is not possible on RBE because it is a containerized environment.

After this CL lands, //:rbe_container_skia_infra will always be built locally, even if Bazel is invoked with --config=remote.

Bug: skia:11111
Change-Id: Ic3fd77fd7a92deb1a85a6a917f4d8e4a2aab5089
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/353516
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Auto-Submit: Leandro Lovisolo <lovisolo@google.com>