[gold] Make diffcalculator ready to be deployed.
This creates a new k8s image and the configuration for it.

I also created an executable called pubsubtool to make
it easier to create pubsub subscriptions/topics from the CLI
and post some test messages. As I tried it out, I revised
some of the topic/subscription conventions from before.

A single data point from testing showed that computing 6
diffs and storing them to the database took about 300ms.
That's a good sign so far, in that doing a little bit of
work takes a little amount of time. It will be interesting
to see how this works at higher scales.

Bug: skia:10582
Change-Id: Ic47a74e6958b65b810c7815a70a29851b43c0ecd
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/365676
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>