[gold] Only compute diff metrics against non-ignored images.
In the skia repo, there is a heavy tail of a few tests that
have thousands of unignored images. Computing a full n by n
comparison would take hours and be mostly useless data anyway.

As such, we only compare against non-ignored images, to decrease
computation. Reminder: our goal is to be able to answer the
question "What is the closest positively or negatively triaged
image to any given image" and the answer to that should only come
from not-ignored traces.

Bug: skia:10582
Change-Id: Icab08916b77f2f5c92f71504d2e47cabf5196953
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/366960
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>