[bazel] Add Infra-PerCommit-Build-Bazel-{RBE,Local} tryjobs.
The local and RBE variants of this tryjob only differ in that the RBE variant passes --config=remote to all Bazel invocations.

I've commented out the code path that adds the --config=remote flag because we have yet to figure out the correct way to grant RBE access to Swarming bots. Both the local and RBE variants of this tryjob will be equivalent until said code is uncommented.

See go/skia-infra-bazel for reference.

Bug: skia:11111
Change-Id: I3f57587f117dcacd9f05808f5b59d8f7ac12102e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/370746
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Eric Boren <borenet@google.com>