Skottie library
Added a new component that allows loading multiple animations at the same time in a zip file.
Moved a method to this helper so edited texts apply to any animation on the library and the main player.
`skottie/modules/skottie-text-editor/text-replace.js `

This is an imported pull request from
https://github.com/google/skia-buildbot/pull/51

GitOrigin-RevId: 8d8462117cd800501c4fba9ddf8ffc3d05f5d86e
Change-Id: I13399289a0972ff043eb671823f65433af1ec7ef
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/373936
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>