[gold] Add support to reading commit_metadata to legacy ingester.
This reads the XML file provided and retrieves the githash of the
tast-tests repo. This will then be able to be seamlessly plugged into
the legacy system so we can see the results while the final bits
of the new system are being developed.

In order to work in prod, we'll need object read access given to
the production service account. I've sent an email to request that.

Bug: skia:11367
Change-Id: Ic7fe839cb41c2c37d5aa2faa309f75861180f419
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/402976
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>