[gold] Make new search API opt-in and fix a few bugs.
If we ingested a second data point for a trace at the same commit
we had observed previously, we were inconsistent about updating
TraceValues and ValuesAtHead. This was noticed because on the infra
repo, we currently have build_system as an optional key. A follow-on
CL will change that to be a normal key.

The new API wasn't filling out the deprecated field
tileSize, making the entire dots-sk not draw. This removes that field
from the frontend and backend.

This adds one new index to help make loading trace history faster.

Bug: skia:10582
Change-Id: I7c983c3914d3ab331b4f272d7effeb30b1d077ed
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/403538
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>