[gold] Add support for public views.
We keep a list of public traces in RAM and use that to make sure
we are only showing data from that.

This is much simpler than trying to filter those out in SQL. If we
did want to use that approach, we would probably want to use more
materialized views.

Bug: skia:10582
Change-Id: Id4b8b3032f40625c3a41ec972e9d4e57f6488b76
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/405261
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>