[gold] Implement status endpoint (and caching system)
This should replace //golden/go/status.

One thing that it does *not* do that the old system did
was refresh counts when something was triaged; instead it
updates the cache on a fixed interval.

Like the previous implementation, we use a cache because
this RPC is requested very frequently and needs to do joins on
tables/indexes which are rather large.

Change-Id: Ibdc7b22251a375a61e639dc799b0ba67c21c18b1
Bug: skia:10582
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/420656
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>