[gold] Implement cl redirect in SQL
I tested the SQL query on an arbitrary CL on the Skia instance
to help optimize it for speed and validate correctness.

Unlike some other RPCs, this endpoint is not versioned, so it
would be more difficult/disruptive to add in a v2. Therefore,
we are just replacing it in place.

Bug: skia:10582
Change-Id: Ida9f924d005bd6cbc0040aebac48c187d99558a7
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/421318
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>