[gold] Add v2 version of diffworker
This is hopefully more resilient to flaky traces. Before, our
defense against flaky traces was that a human would mark them
as ignored. This does not work out in practice, and flaky traces
could cause the previous version to grind to a halt and get
behind by computing too many diffs.

This way does a quick check to see if we can afford to simply
compute all diffs. If there are too many digests to do so, it
picks a bounded subset of common/recent digests to compare
against one another.

This seems much more adept at handling the load observed for
the Skia instance. It is also simpler, instead of having to
query the DB for a rectangle of DiffMetrics (using many small
queries in parallel), it is able to use a single ANY() query
to see what has been computed. This makes the code faster as
well as simpler by removing the channels feeding into multiple
goroutine workers.

Bug: skia:12221
Change-Id: I3787376af9d271df13abcbab70440a8894da701a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/433838
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>