Build and deploy metadata server to jumphosts via Ansible.
The metadata_server_ansible is an application that servers up
OAuth2 Access Tokens from the URL:

   /computeMetadata/v1/instance/service-accounts/default/token

This service is used by Swarming running on all the test machines
in the lab, and the metadata server is run on all the jumphosts.

The metadata_server_ansible application emulates the functioning
of the metadata server on GCP:

  https://cloud.google.com/compute/docs/metadata/overview

This has been run and all jumphosts are now running metadata
servers:

  https://console.cloud.google.com/logs/viewer?project=skia-public&advancedFilter=logName%3D%22projects%2Fskia-public%2Flogs%2Fmetadata_server_ansible%22

These are currently unused as the "real" metadata servers are
running on different ports. Once these are shown to be reliable
the bots will be switched over to use these metadata servers
and we can remove all the old push and k3s based metadata servers.

Bug: skia:12063
Change-Id: I573f79e40313df3df3a771a46dac3dc4b7ea79a0
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/438659
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Erik Rose <erikrose@google.com>
Reviewed-by: Eric Boren <borenet@google.com>