[infra-sk] Add puppeteer tests for task-driver-sk
While debugging another CL, I noticed some TS compilation
errors for infra-sk's puppeteer tests that were not causing
the task to fail. The errors were because the task-driver-sk
element depended on a file that was located in //task_driver
and hadn't been copied in.

I tried to figure out why that compilation error wasn't causing
the mocha tests from failing and it appears the error was a
complaint about any, and the TS compiler was still able to
emit JS. As such, the mocha tests could run and function, so
there was no failure.

I've fixed the file, added puppeteer tests for that element
and changed -r ts-node/register to the longer form
--require so it's a little easier to research in the future.

Change-Id: I12433daaae412b7812759cc76811a7ebbb25a345
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/441362
Reviewed-by: Eric Boren <borenet@google.com>