[pushk] Instead of commiting directly, use rubber stamper
This does the equivalent of
git cl upload --bypass-hooks --send-mail --enable-auto-submit
  --reviewers=rubber-stamper@appspot.gserviceaccount.com --force

but without requiring a dependency on depot_tools.

Empirical testing led us to deduce that a random Change-Id
works and we don't need to compute it the same way that the
presubmit hook does.

Change-Id: I5068a7492aced0c79695d205feac1c63b3e8f033
Bug: skia:12124
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/442273
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>