[machines] adb reconnect if an Android device goes offline.
Android device can go 'offline', but running 'adb reconnect offline' may
reconnect them, so add that as a new step to adb.go.

The added EnsureOnline() checks the status of the connection and tries
to reconnect if the device is offline, otherwise it reports the state
of the device.

Change-Id: Ide7e9dd92951f313c9898f6dec3b4d5ca1206347
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/448466
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>