[gold] Add GCS -> Pub/Sub for tryjob data
While investigating an alert about polling being stalled,
I realized I hadn't turned on the GCS=>Pub/Sub for tryjob
data on the new topic/subscription. We had just been relying
on the backup polling, which worked fine except for Chrome.
Chrome generates many small files and the polling was taking
too long to comb through them all (which is a single-threaded
process).

While fixing that, I updated pubsubtool to log more clearly
(and not spew a bunch of unhelpful goroutine stuff on exit)
and the docs about this type of alert. Additionally, I made
it use the local user's credentials instead of using a JSON
file in GOOGLE_APPLICATION_CREDENTIALS.

I also lowered the threshold on that type of alert to make
us more sensitive to it.

Not seen in this CL was me deleting the old ingestion
topics (e.g. gold-ingestion+gold-skia-eventbus-bt) and
GCS=>Pub/Sub links. I also added these steps to
go/setup-gold-instance

Bug: skia:10582
Change-Id: Ia78ab132733d4ddd2f99a4129d699c04df27759c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/449059
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>