[gold] Address issue with contention blocking diff-calculation
This impacted the Skia instance where new tests were not getting
added to the list of groupings we needed to compute. I believe
the all-in-one transaction was timing out too much.

This splits it up and makes it easier to determine if we should
stop computing diffs for groupings no longer in the current
window.

Change-Id: I6317a316ab4f3c09bfb2c5fe624c20cd5b18b912
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/449178
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>