[machines] Sketch in device-editor-sk
In a follow-on CL, I'll add some boxes to device-editor-sk to
allow us to specify a user and IP address for Chrome OS devices,
as well as the dimensions to include (e.g. the GPU)

I'm trying something new here with having ClearDeviceEvent
as an exported const. I hope this will make it explicit which
components rely on the event, and the typescript compiler
will catch us if we remove it w/ users still around.

Change-Id: I2b96858eb338977a6bd82a2dd8a38d178868e564
Bug: skia:12401
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/449191
Reviewed-by: Joe Gregorio <jcgregorio@google.com>