[gold] Fix backups using AUTH=implicit aka Workload Identity
This change goes hand in hand with https://skia-review.googlesource.com/c/k8s-config/+/451488

We'll need to re-run it when the schema is updated, so the
new tables are properly backed up.

As such, the sqlinit command will remove existing backup
schedules (but not the backup files themselves) before
re-creating them.

A follow-up CL will monitoring to the SHOW SCHEDULES table
so we can avoid being broken for so long without knowing
(two of the backups started to fail because of deleted tables,
one failed because the GCS key in the metadata expired).

Change-Id: I7b3b6217b69da0376b41e092e650492d8f775c7d
Bug: skia:12461, skia:12320
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/451557
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>