[bazel] Use bazelisk on all CI tasks.
This will make it easier for us to upgrade to a new Bazel version, because we won't have to upload a new version of the skia/bots/bazel CIPD package when a new version of Bazel is released. Instead, it will suffice to edit //.bazelversion as needed.

On the Infra-PerCommit-*-Bazel-* CI tasks, Bazelisk will download the Bazel version specified in //.bazelversion on the first "bazelisk" invocation. This download is ~50MB, which makes tasks take an additional couple of seconds to complete. This extra latency is probably worth the convenience.

Bug: skia:11111
Change-Id: Ifafe91dd0ead8c6c81a382fba126009fb3a608db
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/479558
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>