[gazelle] Do not warn about css files being included in sass_library
This does not autogenerate anything new, but it at least silences
the warnings.

What we probably want to do is stop ignoring .css files in parsers.go
and generate rules for those (we need to wrap them in a sass_library
first). This should allow npm includes to be handled automatically.

Change-Id: I8e0c98a30c2961a176f2f1e303eccfe0383db09e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/488336
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>