[skottie] Port to Bazel
This updates the docs html to load canvaskit.js separately
from inline-bundle.js (and removes the useless inline-bundle.css).

I think I've finally settled on a way to handle the use of
canvaskit binaries on the CI vs being deployed. There is a
folder wasm_libs that fetches the js/wasm from the docker container
via genrules, and that is used in most places. When we build
our Docker container to go on k8s, we just include the
canvaskit js/wasm from ./build (which have placeholders
checked in).

This needed to rename skottie-text-editor to skottie-text-editor-sk
in order to make the Gazelle generated rules properly pass
around the css rules.

Change-Id: Ia6d731f70caea01fabed34a495ca237a218794c4
Bug: skia:10614
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/487224
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>