Change jsfiddle, particles, shaders to use CanvasKit like skottie
I like the approach where we use a wasm_libs subfolder with
genrules to get the canvaskit.js/wasm/etc libraries and when
we push, we download them ourselves. This makes it easier for
tests to use a build of canvaskit and have type-checked symbols.

This CL structures all of these apps to be the same.
Debugger (as of yet unmigrated to bazel) will also follow this
structure.

Change-Id: I4eb8209745ff4b1d55d228e7b5e4489cacd0e7d4
Bug: skia:10614
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/488526
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>