Factor out SkSL-related constant strings to a separate lib.
This will allow us to share CodeMirror configuration between the
debugger and shader playground.

The implementation of `words` changed slightly to more closely match the
`words` function for other languages. (Specifically, replaced
`backticks` with 'foo ' + 'bar' so that we can split without needing a
regular expression or needing to special-case empty strings. At one
point I also managed to get it down to a one-liner with `reduce`, but it
was more difficult to understand, so I undid that.)

Bug: skia:12778
Change-Id: Icf9ff70b7c799b8eaeaa8fa2823ec17114a3088b
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/491277
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>