[perf] Add a sort history to pivot tables.
This enables better sorting behavior, i.e. when you click on
col A to sort, then on col B to sort, if there are ties in
col B they are broken by the existing order in col A, just
like you would get when sorting by columns in a spreadsheet.

In a future CL history will be serialized and included in the Perf URL
so that when people share Perf links they always see the table in the
exact same way.

This is not technically 'stable sort', while each sort action by
the user looks like it is doing a stable sort, which is the goal, we
are really doing an absolute sort based on a memory of all previous
sort actions.

Change-Id: Iec0e4aa9e91a8d9c58d17ad50c752c12a1027db6
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/491396
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>