Allow debugger to load a trace from Local Storage.
This gives shaders.skia.org a simple pathway to hand off a trace to the
debugger. When a trace is collected, we can store it in Local Storage
and then redirect the user to `/debug?local-storage`, which will detect
a usable DebugTrace JSON in Local Storage and then immediately debug it
instead of prompting the user for a file.

The query parameter exists to avoid accidentally locking users into a
trace forever. Traces can always be loaded by dragging them in, but
without the initial prompt, a user won't know to try it.

Testing is accomplished by giving the test code hooks which can replace
the window's query parameter and Local Storage object at
initialization time.

Bug: skia:12818
Change-Id: I8f4a61a9efd8f523a2e517d40f6e5b42d3753fae
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/494823
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>